# Speed up process startup time by avoiding importing packages on version check