`importlib.metadata` can check the version of a package without importing it. That makes our nanny and worker processes come up **much** faster, especially considering that the Nanny doesn't even have any usage for this (i.e. this also saves a bit on memory but that should be marginal in _most_ cases).

Launching a LocalCluster with 10 workers on my machine is 2-3x faster with this change.

```
# Main
▶ time python -c "import distributed; distributed.LocalCluster(n_workers=10)"
10.98s user 3.07s system 147% cpu 9.516 total
▶ time python -c "import distributed; distributed.LocalCluster(n_workers=10)"
10.31s user 2.67s system 158% cpu 8.166 total

# This PR
▶ time python -c "import distributed; distributed.LocalCluster(n_workers=10)"
6.73s user 1.89s system 216% cpu 3.976 total
▶ time python -c "import distributed; distributed.LocalCluster(n_workers=10)"
6.52s user 1.76s system 252% cpu 3.275 total
▶ time python -c "import distributed; distributed.LocalCluster(n_workers=10)"
6.61s user 1.79s system 268% cpu 3.125 total
```

I hope this helps speed up our CI infrastructure and helps us avoid a couple of flaky tests which are related to Nanny processes startup failures