@mrocklin 2025-04-16
That's awesome.

On Wed, Apr 16, 2025 at 8:15 AM Florian Jetter ***@***.***>
wrote:

> importlib.metadata can check the version of a package without importing
> it. That makes our nanny and worker processes come up *much* faster,
> especially considering that the Nanny doesn't even have any usage for this
> (i.e. this also saves a bit on memory but that should be marginal in
> *most* cases).
>
> Launching a LocalCluster with 10 workers on my machine is 2-3x faster with
> this change.
>
> # Main
> ▶ time python -c "import distributed; distributed.LocalCluster(n_workers=10)"
> 10.98s user 3.07s system 147% cpu 9.516 total
> ▶ time python -c "import distributed; distributed.LocalCluster(n_workers=10)"
> 10.31s user 2.67s system 158% cpu 8.166 total
>
> # This PR
> ▶ time python -c "import distributed; distributed.LocalCluster(n_workers=10)"
> 6.73s user 1.89s system 216% cpu 3.976 total
> ▶ time python -c "import distributed; distributed.LocalCluster(n_workers=10)"
> 6.52s user 1.76s system 252% cpu 3.275 total
> ▶ time python -c "import distributed; distributed.LocalCluster(n_workers=10)"
> 6.61s user 1.79s system 268% cpu 3.125 total
>
> I hope this helps speed up our CI infrastructure and helps us avoid a
> couple of flaky tests which are related to Nanny processes startup failures
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/dask/distributed/pull/9048
> Commit Summary
>
>    - a1267e6
>    <https://github.com/dask/distributed/pull/9048/commits/a1267e65fce609957caa0a811cf11d65e4896208>
>    Avoid importing critical packages on process startup
>    - ff1a53d
>    <https://github.com/dask/distributed/pull/9048/commits/ff1a53d29347db886af147e59dc3b9d3b7dd93ac>
>    Fix test
>
> File Changes
>
> (2 files <https://github.com/dask/distributed/pull/9048/files>)
>
>    - *M* distributed/tests/test_versions.py
>    <https://github.com/dask/distributed/pull/9048/files#diff-4553904db2be9e5ecf54a3405fea89ada4466eabd642a4e9fa893a66f9ae495c>
>    (11)
>    - *M* distributed/versions.py
>    <https://github.com/dask/distributed/pull/9048/files#diff-86e7e03bee50d99539a98a0acdf97d2ea0c177442d9d5e174630d8a6c5349a3f>
>    (48)
>
> Patch Links:
>
>    - https://github.com/dask/distributed/pull/9048.patch
>    - https://github.com/dask/distributed/pull/9048.diff
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/dask/distributed/pull/9048>, or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AACKZTFGIN6YZFFACJBPZTT2ZZJVTAVCNFSM6AAAAAB3IIP2HOVHI2DSMVQWIX3LMV43ASLTON2WKOZSHE4TSNRSGY3DIMY>
> .
> You are receiving this because you are subscribed to this thread.Message
> ID: ***@***.***>
>