@fjetter 2025-04-16
curiously, CI performance is much worse 🤔