@fjetter 2025-04-16
Well, `importlib.metadata.version` *is* much faster than an initial import but it is much slower compared to an import that already happened. Therefore, all of our test that do not use dedicated processes got slower... that should be fixed with a simple cache

Test runtime on my machine for test_scheduler.py

main: ~60s
PR w/out cache: ~120s
PR w/ cache: ~45s

Let's see if this also translate to CI